pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-memory-hybrid"
version = "0.4.0"
description = "A semantic memory system for AI agents combining OpenViking AFS and LightMem async processing"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
authors = [
    {name = "Lilith Liu", email = "your.email@example.com"}
]
keywords = ["agent", "memory", "llm", "ai", "rag", "embeddings", "openviking", "lightmem"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Lilith6666liu/Agent-memory-hybrid"
Documentation = "https://github.com/Lilith6666liu/Agent-memory-hybrid#readme"
Repository = "https://github.com/Lilith6666liu/Agent-memory-hybrid"
Issues = "https://github.com/Lilith6666liu/Agent-memory-hybrid/issues"

[tool.hatch.build.targets.wheel]
packages = ["memory"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
